move DateTimer to package datetimer
