+++ src/main/java/com/jarq/system/helpers/datetimer/IDateTimer.java
@@ -1,4 +1,4 @@
-package com.jarq.system.helpers;
+package com.jarq.system.helpers.datetimer;
 
 public interface IDateTimer {
